python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-cost-calculator"
version = "0.2.0"
description = "Python SDK to calculate LLM API costs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "jsonpath-ng>=1.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "python-dotenv>=1.0.0",
]

[tool.setuptools.package-data]
ai_cost_calculator = ["data/*.json"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

